+++ packages/css/src/components/description/description.template.ts
@@ -4,12 +4,12 @@ import { classMap } from 'lit-html/directives/class-map.js';
 import { ifDefined } from 'lit-html/directives/if-defined.js';
 import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
 
-export function descriptionTemplate({ id, open, term, content, type }: Description) {
+export function descriptionTemplate({ id, open, term, content }: Description) {
   return html`
     <a
       id="${id}-term"
       href="#${id}-content"
-      class="${classMap({ [`dso-description-${type}`]: true, 'dso-open': open })}"
+      class="${classMap({ [`dso-description-term`]: true, 'dso-open': open })}"
       aria-expanded="${ifDefined(open ? 'true' : 'false')}"
       aria-controls="${id}-content"
     >${unsafeHTML(term)}</a>