+++ packages/sources/src/components/description/description.models.ts
@@ -1,5 +1,4 @@
 export interface Description {
-  type: 'term' | 'note';
   id: string;
   term: string;
   content: string;